pyproject.toml
[project]
name = "math-logic-agent"
version = "0.1.0"
description = "Hybrid local+cloud math logic sidecar for Copilot workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "pypdf>=5.2.0",
    "cryptography>=43.0.0",
    "python-pptx>=1.0.2",
    "scikit-learn>=1.5.2",
    "sympy>=1.13.3",
    "python-dotenv>=1.0.1",
    "typer>=0.15.1",
    "rich>=13.9.4",
    "openai>=1.59.7",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
]
ocr = [
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
]

[project.scripts]
mla = "math_logic_agent.cli:app"

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
